+++ src/Statistics/NanQuantile.jl
@@ -2,96 +2,29 @@ import Statistics: quantile!
 using Base: _unsafe_getindex!, _unsafe_setindex!, Slice
 
 include("quantile_base.jl")
-
-
-function NanQuantile_low!(R::AbstractArray{<:Real,N}, A::AbstractArray{<:Real,N};
-  probs::Vector=[0, 0.25, 0.5, 0.75, 1], dims::Integer=N, na_rm::Bool=true) where {N}
-
-  fun = na_rm ? _nanquantile! : quantile!
-  itershape = ntuple(d -> d in dims ? Base.OneTo(1) : axes(A, d), ndims(A))
-  inds = CartesianIndices(itershape)
-
-  dim_mask = ntuple(d -> d in dims, ndims(A))
-
-  # intermediate variables
-  ntime = size(A, dims)
-  nprob = length(probs)
-  Aslice = zeros(eltype(A), ntime)
-  r = zeros(eltype(A), nprob)
-
-  _Slice_A = Slice.(axes(A))
-  _Slice_R = Slice.(axes(R))
-
-  # idx = Vector{Any}(nothing, ndims(A))
-  # ridx = Vector{Any}(nothing, ndims(R))
-  @inbounds for I in inds
-    # for k in eachindex(dim_mask)
-    #   idx[k] = dim_mask[k] ? _Slice_A[k] : I[k]
-    #   ridx[k] = dim_mask[k] ? _Slice_R[k] : I[k]
-    # end
-    idx = ifelse.(dim_mask, _Slice_A, Tuple(I))
-    ridx = ifelse.(dim_mask, _Slice_R, Tuple(I))
-    _unsafe_getindex!(Aslice, A, idx...)
-    fun(r, Aslice, probs)
-    R[ridx...] = r
-  end
-  R
-end
-
-
-"""
-  $(TYPEDSIGNATURES)
-
-# Examples
-```julia
-using Test
-
-dates = make_date(2010, 1, 1):Day(1):make_date(2010, 12, 31)
-ntime = length(dates)
-arr = rand(Float32, 140, 80, ntime)
-arr2 = copy(arr)
-
-# default `na_rm=true`
-@test NanQuantile([1, 2, 3, NaN]; probs=[0.5, 0.9], dims=1) == [2.0, 2.8]
-
-@time r0 = _nanquantile(arr, dims=3) # low version
-@time r2_0 = NanQuantile_low(arr; dims=3, na_rm=false)
-@time r2_1 = NanQuantile_low(arr; dims=3, na_rm=true)
-
-@test r2_0 == r2_1
-@test r2_0 == 20
-@test arr2 == arr
-```
-
-!!!`NanQuantile_low(na_rm=rue)` is 3~4 times faster than `_nanquantile(na_rm=true)`
-"""
-function NanQuantile_low(A::AbstractArray{T,N};
-  probs::Vector=[0, 0.25, 0.5, 0.75, 1], dims::Integer=N, na_rm::Bool=true, dtype=nothing) where {T<:Real,N}
-
-  dtype = dtype === nothing ? eltype(A) : dtype
-  Size = size(A) |> collect
-  Size[dims] = length(probs)
-  R = zeros(dtype, Size...)
-  NanQuantile_low!(R, A; probs, dims, na_rm)
-end
-
+include("NanQuantile_low.jl")
 
 # `NanQuantile`与`NanQuantile_low`性能相当
 function NanQuantile(x::AbstractArray{T,N};
   probs=[0, 0.25, 0.5, 0.75, 1], dims::Integer=N,
-  na_rm::Bool=true, dtype=nothing) where {T<:Real, N}
+  na_rm::Bool=true, dtype=nothing, use_zi=true) where {T<:Real, N}
 
   dtype = dtype === nothing ? T : dtype
   fun! = na_rm ? _nanquantile! : quantile!
-  ntime = size(x, dims)
-  nprob = length(probs)
-  zi = zeros(eltype(x), ntime)
-  qi = zeros(dtype, nprob)
+  
+  qi = zeros(dtype, length(probs))
 
-  mapslices(xi -> begin
-      copyto!(zi, xi)
-      fun!(qi, zi, probs)
-    end, x; dims=dims)
+  if use_zi
+    ntime = size(x, dims)
+    zi = zeros(eltype(x), ntime)
+    
+    mapslices(xi -> begin
+        copyto!(zi, xi)
+        fun!(qi, zi, probs)
+      end, x; dims=dims)
+  else
+    mapslices(xi -> fun!(qi, xi, probs), x; dims=dims)
+  end
 end
 
 function NanQuantile(x::AbstractArray{T,N}, probs;