+++ src/Statistics/Statistics.jl
@@ -9,5 +9,5 @@ include("NanQuantile.jl")
 
 export mean, median, quantile, movmean, weighted_mean, weighted_movmean
 export apply
-export _nanquantile!, nanquantile, NanQuantile,NanQuantile_low, NanQuantile_low!
+export _nanquantile!, nanquantile, NanQuantile
 export lm, linreg, linreg_fast, linreg_simple