+++ News.md
@@ -1,3 +1,17 @@
+# v0.1.9
+
+- fix `cal_mTRS_base` when `q` is not updated when `method_q == "mapslices"`
+
+- improve `cal_mTRS_base` performance by `filter_mds` and `format_md`, about 3-5 times faster
+
+- deprecated `NanQuantile_low`
+
+
+# v0.1.8 
+
+- major updates: `NaNQuantile`统一命名规则
+
+
 # v0.1.4 (2022-08-25)
 
 - The best version NanQuantile_low
@@ -5,7 +19,3 @@
 - fix `melt_list`
 - fix `dt_merge` when `by` is a Symbol
 - rm `write_fig`
-
-# v0.1.8 
-
-- major updates: `NaNQuantile`统一命名规则